fixed a problem with the way we serialized tokens ...
before, we only overwrote the stored value (e.g. `token.form`) if it was `!== undefined && !== '')`

but this could present problems re the way we deserialize strings

for example, if the value of `token.form` was set to "\t", then this would serialize out as "\t", not "_".  when we attempt to deserialize the serialized token, we will end up splitting on that char, even though it should be interpreted as content.

as a result, we get an unstable edge condition on serializing/deserializing data

i've changed the behavior to now reject the stored value (`token.<key>`) and serialize the default ("_") if the stored value is only whitespace

basically, this means i've changed all the `this.<key> !== ''` to `!/^[ \t]*$/.test(this.<key>)`